% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{gen_data_db}
\alias{gen_data_db}
\title{Generate full db from a Cookbook of Recipes}
\usage{
gen_data_db(
  cbook,
  db = list(),
  ns = setNames(rep(500, NROW(cbook)), cbook$table)
)
}
\arguments{
\item{cbook}{tribble. Columns: table, scaff_ref, data_rec, na_rec}

\item{db}{list. Named list of already existing tables}

\item{ns}{named numeric. Ns for use when generating independent tables.}
}
\value{
A named list of tables of generated data.
}
\description{
Generate full db from a Cookbook of Recipes
}
