% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\name{init_new_cols}
\alias{init_new_cols}
\title{Initialize new columns of the correct length}
\usage{
init_new_cols(
  n,
  colnames = names(colclasses),
  colclasses = setNames(rep(NA, length(colnames)), colnames)
)
}
\arguments{
\item{n}{numeric(1). The length (number of rows) to use when initializing.}

\item{colnames}{character. Vector of column names to use. Can be omitted if \code{colclasses} is specified.}

\item{colclasses}{named character. Optional. Names must be identical to \code{colnames} if specified, values are classes such that \code{as(NA, .)} will succeed. Defaults to \code{NA} for each column, indicating character columns.}
}
\value{
A data.frame with the new columns and \code{n} rows.
}
\description{
Initialize new columns of the correct length
}
\examples{

init_new_cols(5, c("col1", "col2"))

init_new_cols(5, colclasses = c(col1 = NA, col2 = "integer"))

}
