% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{inject_nas}
\alias{inject_nas}
\title{Apply random or systematic missingness to existing data according to recipe}
\usage{
inject_nas(tbl, recipe)
}
\arguments{
\item{tbl}{data.frame/tibble. The already generated data to inject missingness into}

\item{recipe}{tibble. A recipe for generating missingness positions}
}
\value{
\code{tbl}, with missingness injected into it as laid out by \code{recipe}
}
\description{
Apply random or systematic missingness to existing data according to recipe
}
\details{
Unlike in data-generation recipes, the \code{func} column in a missingness recipe must return a logical vector of length equal to what it receives as the \code{.df} argument. \code{TRUE} in the return value indicates that position in the column being processed should be set to missing (\code{NA}), while \code{FALSE} indicates the value already there should remain unchanged.
}
\examples{
library(tibble)
missrec <- tibble(variables = "wt", func = list(function(.df) rep(c(TRUE, FALSE), times = c(3, NROW(.df) - 3))), func_args = list(NULL))
newdat <- inject_nas(mtcars, missrec)
head(newdat)

}
