% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lookup_fun}
\alias{lookup_fun}
\title{Lookup function from value you recipe column}
\usage{
lookup_fun(str)
}
\arguments{
\item{str}{ANY. A function (immediately returned) or a character(1) value of the form \code{func},
\code{"pkg::func"} or \code{"pkg:::func"}. Any other value will result in an error.}
}
\value{
A function
}
\description{
This function is an internal utility exported for its usefulness
when debugging recipes. Normal workflows will not involve calling
it directly.
}
\examples{
lookup_fun(rnorm)
lookup_fun("rnorm")
lookup_fun("stats::rnorm")
lookup_fun("stats:::rnorm")
}
