% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_utils.R
\name{miss_as_block}
\alias{miss_as_block}
\title{Function constructor for injecting "missing together" NAs to columns}
\usage{
miss_as_block(p, nvars)
}
\arguments{
\item{p}{numeric(1). Proportion of observations to change to missing}

\item{nvars}{numeric(1). Number of columns this behavior should be applied to.}
}
\value{
a function suitable for use in missingness recipe, which when given \code{.df}, the data
will return an N x nvars logical matrix.
}
\description{
This constructor creates a function which will set all \code{nvars} columns
to NA together for rows selected for missing data.
}
