% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_utils.R
\name{miss_at_random}
\alias{miss_at_random}
\title{Function constructor for injecting independent missing-at-random NAs to columns}
\usage{
miss_at_random(p, nvars)
}
\arguments{
\item{p}{numeric(1). Proportion of observations to change to missing}

\item{nvars}{numeric(1). Number of columns this behavior should be applied to.}
}
\value{
a function suitable for use in missingness recipe, which when given \code{.df}, the data
will return an N x nvars logical matrix.
}
\description{
Function constructor for injecting independent missing-at-random NAs to columns
}
