% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\docType{data}
\name{no_deps}
\alias{no_deps}
\title{Sentry values for Recipe construction}
\format{
An object of class \code{character} of length 0.
}
\usage{
no_deps
}
\description{
Sentry values for Recipe construction
}
\details{
\code{no_deps} indicates no dependencies
}
\keyword{datasets}
