% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\docType{data}
\name{pct_orig}
\alias{pct_orig}
\alias{secs_per_year}
\alias{secs_per_day}
\alias{rand_posixct}
\title{Generate a sample of random datetimes}
\format{
An object of class \code{character} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
pct_orig

secs_per_year

secs_per_day

rand_posixct(
  start,
  end,
  max_duration_secs = NULL,
  multiplier = if (is(start, "Date")) secs_per_day else 1,
  n = max(length(start), length(end))
)
}
\arguments{
\item{start}{POSIXct or Date. Earliest possible datetime for thte sample}

\item{end}{POSIXct or Date. latest possible datetime for thte sample}

\item{max_duration_secs}{numeric. Number of seconds to use to generate alernate end if \code{end} has a missing value.}

\item{multiplier}{numeric. Used internally.}

\item{n}{numeric. Length of sample. Default to max of \code{length(start)} and \code{length(end)}.}
}
\value{
A \code{POSIXct} vector of datetimes.
}
\description{
Generates random datetimes that are, elementwise, between \code{start} and \code{end}
}
\examples{
rand_posixct("2020-01-01", "2021-01-01", n = 2)
rand_posixct(c("1995-04-01", "2000-01-01"), c("2000-04-01", "2000-01-30"))

}
\keyword{datasets}
