% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\name{rand_posixct}
\alias{rand_posixct}
\title{Generate a sample of random datetimes}
\usage{
rand_posixct(
  start,
  end,
  max_duration_secs = NULL,
  multiplier = if (is(start, "Date")) secs_per_day else 1,
  n = max(length(start), length(end))
)
}
\arguments{
\item{start}{POSIXct or Date. Earliest possible datetime for thte sample}

\item{end}{POSIXct or Date. latest possible datetime for thte sample}

\item{max_duration_secs}{numeric. Number of seconds to use to generate alernate end if \code{end} has a missing value.}

\item{multiplier}{numeric. Used internally.}

\item{n}{numeric. Length of sample. Default to max of \code{length(start)} and \code{length(end)}.}
}
\value{
A \code{POSIXct} vector of datetimes.
}
\description{
Generates random datetimes that are, elementwise, between \code{start} and \code{end}
}
