% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\name{sample_fct}
\alias{sample_fct}
\alias{sample_yn}
\alias{rep_n}
\alias{seq_n}
\title{Create a factor with random elements of x}
\usage{
sample_fct(x, n, ...)

sample_yn(n)

rep_n(val, n, ...)

seq_n(n, ...)
}
\arguments{
\item{x}{character vector or factor, if character vector then it is also used
as levels of the returned factor, otherwise if it is a factor then the
levels get used as the new levels}

\item{n}{number of observations to sample.}

\item{...}{arguments passed on to \code{\link{sample}}}

\item{val}{ANY. Single value to be repeated n times}
}
\value{
a factor of length N
}
\description{
Sample elements from x with replacing and build a factor
}
\examples{
sample_fct(letters[1:3], 10)
sample_fct(iris$Species, 10)

rep_n("aaa", 5)

}
