% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_recipe_utils.R
\name{subjid_func}
\alias{subjid_func}
\title{Generate sequence of "subject id"s}
\usage{
subjid_func(n, prefix = "id", suffix = NULL, sep = "-")
}
\arguments{
\item{n}{numeric(1). number of ids to generate}

\item{prefix}{character(1). Prefix to prepend to the generated numeric ids. Defaults to \code{"id"}}

\item{suffix}{character(1). Suffix to append to generated ids. Defautls to \code{NULL} (no suffix).}

\item{sep}{character(1). String to use as separator when combining \code{prefix}, number, and \code{suffix}.}
}
\value{
sequence from 1 to \code{n}, prepended with \code{prefix}, and appended with \code{suffix}, separated by \code{sep}
}
\description{
Generate sequence of "subject id"s
}
\examples{
subjid_func(5)
}
