% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_recipe.R
\name{validate_recipe_deps}
\alias{validate_recipe_deps}
\title{Validate recipe for circular dependencies}
\usage{
validate_recipe_deps(recipe, seed_df = NULL)
}
\arguments{
\item{recipe}{tibble. Table recipe.}

\item{seed_df}{tibble. Seed/pre-existing data.frame or NULL.}
}
\value{
\code{TRUE} if successful, throws an error if not
}
\description{
Validate recipe for circular dependencies
}
